extract account creator and config fetching from main service class
